lpc/requirements.txt
Django>=1.9,<1.10
django-shop>=0.9,<1.0
djangocms_text_ckeditor>=2.9,<2.10
djangocms-cascade>=0.8,<0.9
djangocms-bootstrap3>=0.1,<0.2
django-admin-sortable2>=0.6,<0.7
djangorestframework>=3.1,<3.2
django-rest-auth>=0.5,<0.6
django-fsm>=2.3,<2.4
django-fsm-admin>=1.2,<1.3
django-libsass>=0.4,<0.5
django-filter>=0.13,<0.14
django-parler>=1.6,<1.7
djangoshop-stripe>=0.1,<0.2
